requirements.txt
Django==2.2.4
django-common-helpers==0.9.2
django-cors-headers==3.1.0
django-crontab==0.7.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.0.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
